requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
bootstrap4==0.1.0
certifi==2020.6.20
click==7.1.2
Django==2.2.14
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
importlib-metadata==1.7.0
pathspec==0.8.0
psycopg2-binary==2.8.5
pytz==2020.1
regex==2020.7.14
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
zipp==3.1.0
